+++ app/src/main/java/com/e/baize/Player.java
@@ -8,20 +8,20 @@ import java.util.List;
 
 public class Player extends AppCompatActivity implements Serializable  {
     public String Name;
-    public List<Integer> ScoreHistory = new ArrayList<>();
+    public List<Float> ScoreHistory = new ArrayList<>();
     private int FrameCount;
 
-    public void setScore(int score) {
+    public void setScore(float score) {
         ScoreHistory.add(score);
     }
 
-    public List<Integer> getScoreHistory() {
+    public List<Float> getScoreHistory() {
         return ScoreHistory;
     }
 
     public int getScore() {
         int score = 0;
-        for (int i : ScoreHistory) {
+        for (Float i : ScoreHistory) {
             score += i;
             }
         return score;
@@ -48,7 +48,7 @@ public class Player extends AppCompatActivity implements Serializable  {
     public List<String> getFrameScores() {
         List<String> scoreList = new ArrayList<>();
         for (int i = 0; i < ScoreHistory.size(); i++) {
-            String sScore = (Integer.toString(ScoreHistory.get(i)));
+            String sScore = (Float.toString(ScoreHistory.get(i)));
             scoreList.add(sScore);
         }
         return scoreList;
